Move algorithms to ftensp
