+++ include/fractool/ftensp/generate_mandelbrot.hpp
