+++ src/CMakeLists.txt
@@ -1,2 +1,3 @@
 add_subdirectory(ftcore)
+add_subdirectory(ftensp)
 add_subdirectory(fractool)
