+++ src/ftensp/CMakeLists.txt
@@ -1,0 +1,5 @@
+add_library(ftensp
+	generate_julia.cpp
+	generate_mandelbrot.cpp)
+target_link_libraries(ftensp PUBLIC ftcore)
+install(TARGETS ftensp LIBRARY)
